+++ src/alter-node.ts
@@ -1,29 +1,56 @@
-import { HtmlAttributesDictionary, HTMLNode } from 'react-native-render-html'
+import { HtmlAttributesDictionary, HTMLNode, HTMLTagNode } from 'react-native-render-html'
 import forEachObjIndexed from 'ramda/es/forEachObjIndexed'
 import strigifyEntities from 'stringify-entities'
 
 function renderOpeningTag(tag: string, attributes: HtmlAttributesDictionary) {
   const strAttributes: string[] = []
-  forEachObjIndexed((attrVal: string, attrKey: string) => {
-    strAttributes.push(`${attrKey}="${strigifyEntities(attrVal)}"`)
+  forEachObjIndexed((attrVal: string|number, attrKey: string) => {
+    strAttributes.push(`${attrKey}="${strigifyEntities(`${attrVal}`)}"`)
   })(attributes)
   return `<${tag}${strAttributes.length ? ' ' : ''}${strAttributes.join(' ')}>`
 }
 
-export function domToHTML(root: HTMLNode): string {
+export interface TableStats {
+  rows: number
+  columns: number
+  characters: number
+}
+
+export function domToHTML(root: HTMLNode, stats: TableStats): string {
+  let html = ''
   if (root.type === 'tag') {
-    const strChildren = root.children.reduce((prev: string, curr: HTMLNode) => `${prev}${domToHTML(curr)}`, '')
-    return `${renderOpeningTag(root.name, root.attribs)}${strChildren}</${root.name}>`
+    const strChildren = root.children.reduce((prev: string, curr: HTMLNode) => {
+      if (curr.type === 'tag' && curr.name === 'tr') {
+        stats.rows += 1
+      }
+      if (curr.type === 'tag' && (curr.name === 'td' || curr.name === 'th') && stats.rows === 1) {
+        stats.columns += 1
+      }
+      return `${prev}${domToHTML(curr, stats)}`
+    }, '')
+    html = `${renderOpeningTag(root.name, root.attribs)}${strChildren}</${root.name}>`
+  } else if (root.type === 'text') {
+    const text = strigifyEntities(root.data)
+    html = text
+    stats.characters += text.length
   }
-  if (root.type === 'text') {
-    return strigifyEntities(root.data)
-  }
-  return ''
+  return html
 }
 
 export default function alterNode(node: HTMLNode) {
   if (node.type === 'tag' && node.name === 'table') {
+    const stats = {
+      html: '',
+      rows: 0,
+      columns: 0,
+      characters: 0
+    }
     node.attribs.cellspacing = '0'
-    node.attribs.rawHtml = domToHTML(node)
+    const html = domToHTML(node, stats)
+    const { rows, columns, characters } = stats
+    node.attribs._rawHtml = html
+    node.attribs._numOfRows = rows
+    node.attribs._numOfColumns = columns
+    node.attribs._numOfChars = characters
   }
 }