+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## v0.3.0
+## v0.3.1
 
 - add `selectableText` field in `tableStyleSpecs` config option
+- export `cssRulesFromSpecs` and `defaultTableStylesSpecs`
+- add FAQ to readme
+- added `numOfRows`, `numOfColumns` and `numOfChars` to `HTMLTable` props, which can help trigger rendering logic depending on table stats
+- export `makeCustomTableRenderer` to customize table component
+- added a new example with custom complex table